+++ jopa-impl/src/test/java/cz/cvut/kbss/jopa/adapters/IndirectMultilingualStringTest.java
@@ -111,4 +111,9 @@ class IndirectMultilingualStringTest {
     void getReferencedStringReturnsWrappedInstance() {
         assertSame(referencedString, sut.unwrap());
     }
+
+    @Test
+    void toStringInvokesToStringOfReferencedInstance() {
+        assertEquals(referencedString.toString(), sut.toString());
+    }
 }