+++ jopa-impl/src/test/java/cz/cvut/kbss/jopa/model/metamodel/ClassFieldMetamodelProcessorTest.java
@@ -61,7 +61,7 @@ class ClassFieldMetamodelProcessorTest {
         final EntityTypeImpl<OWLClassJ> etMock = mock(EntityTypeImpl.class);
         when(etMock.getJavaType()).thenReturn(OWLClassJ.class);
         final TypeBuilderContext<OWLClassJ> context = new TypeBuilderContext<>(etMock, new NamespaceResolver());
-        context.setConverterResolver(new ConverterResolver(new Converters()));
+        context.setConverterResolver(new ConverterResolver(new Converters(new Configuration())));
         final ClassFieldMetamodelProcessor<OWLClassJ> processor = new ClassFieldMetamodelProcessor<>(context,
                 metamodelBuilder);
         final Field field = OWLClassJ.getOwlClassAField();